+++ src/riscv64/mod.rs
@@ -1,0 +1,9 @@
+// Copyright 2024 © Institute of Software, CAS. All rights reserved.
+// Copyright 2019 Amazon.com, Inc. or its affiliates. All Rights Reserved.
+// SPDX-License-Identifier: Apache-2.0
+
+#[allow(clippy::all)]
+#[allow(clippy::undocumented_unsafe_blocks)]
+pub mod bindings;
+
+pub use self::bindings::*;